requirements.txt
numpy
pandas
tweepy
python-dotenv
"apache-airflow[celery]==2.5.0" --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.5.0/constraints-3.7.txt"
